rename and move dir
